requirements.txt
python-dotenv
python-telegram-bot>=20.0
google-genai
supabase
flask
gunicorn